[2주차] 카드 게임 UI적용
