+++ lib/screen/asset_image_name.dart
@@ -1,7 +1,11 @@
 class AssetImageName {
   static String get apple => 'images/apple.jpg';
+
   static String get banana => 'images/banana.jpg';
+
   static String get kiwi => 'images/kiwi.jpg';
+
   static String get orange => 'images/orange.jpg';
+
   static String get strawberry => 'images/strawberry.jpg';
 }
